aaaa

+++ src/modules/quiz/application/useCases/delete.question.use-case.ts
@@ -1,5 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { QuizRepository } from '../../infrastructure';
+import { QuizQueryRepository, QuizRepository } from '../../infrastructure';
+import { NotFoundException } from '@nestjs/common';
 
 export class DeleteQuestionCommand {
   constructor(public id: string) {}
@@ -9,8 +10,15 @@ export class DeleteQuestionCommand {
 export class DeleteQuestionUseCase
   implements ICommandHandler<DeleteQuestionCommand>
 {
-  constructor(private readonly quizRepository: QuizRepository) {}
+  constructor(
+    private readonly quizRepository: QuizRepository,
+    private readonly quizQueryRepository: QuizQueryRepository,
+  ) {}
   async execute({ id }: DeleteQuestionCommand): Promise<void> {
+    const question = await this.quizQueryRepository.getQuestionById(id);
+    if (!question) {
+      throw new NotFoundException();
+    }
     await this.quizRepository.deleteQuestionById(id);
   }
 }
